+++ app/components/buttons/viewer/RightButton.js
@@ -7,12 +7,13 @@ export default class RightButton extends Component {
 
     render() {    
         return (
-            <TooltipOverlay  component={ props => <StyledButton
+            <TooltipOverlay  component={ props => <span className="d-inline-block" {...props}><StyledButton
                     variant="secondary" 
+                    disabled={this.props.disabled}
                     onClick={() => this.props.click(1)} 
-                    className={`btn btn-default fas fa-arrow-right button`} 
+                    className={`btn btn-default fa fa-arrow-right button`} 
                     {...props}
-                />} 
+                /></span>} 
                 text={"Next Image"}
             />    
         )