+++ app/components/gallery/TopBar.js
@@ -5,6 +5,7 @@ import RefreshButton from '../containers/gallery/RefreshButtonContainer';
 import FilterBar from '../containers/gallery/FilterBarContainer'
 import { StyledTopBarDiv, StyledTopBarHR } from '../CommonStyledComponents'
 import ServerIndicator from '../containers/gallery/ServerIndicatorContainer';
+import OpenViewerButton from '../buttons/gallery/OpenViewerButton';
 
 import styled from 'styled-components';
 
@@ -28,7 +29,10 @@ export default class TopBar extends Component {
                         <input type="checkbox" className="custom-control-input" id="drawRsml" defaultChecked={true} onClick={this.props.toggleArch}/>
                         <label className="custom-control-label" htmlFor="drawRsml">Draw Architecture</label>
                     </this.labelDiv>
-                    <ServerIndicator />
+                    <div style={{ marginLeft: 'auto' }} >
+                        <OpenViewerButton />
+                        <ServerIndicator />
+                    </div>
                 </StyledTopBarDiv>
                 <StyledTopBarHR/>
             </div>