+++ app/components/viewer/TopBar.js
@@ -19,28 +19,31 @@ export default class TopBar extends Component {
         let tag = path ? splitPath.fileName : ""; //Matches the file path into the absolute directory path and file name
         const folderFiles = this.props.files[splitPath.path];
 
-        const noRSMLInFolder = !Object.values(folderFiles).some(file => file.parsedRSML);
+        const noRSMLInFolder = folderFiles && !Object.values(folderFiles).some(file => file.parsedRSML);
+
+        let message = ""
+        if (noRSMLInFolder) 
+            message = <b>"No RSML found in Folder"</b>;
+        else if (folderFiles) 
+            message = <><b>Open Image: </b>{`${Object.keys(folderFiles).indexOf(tag) + 1} of ${Object.keys(folderFiles).length}`}</>
+
         return (
             <div>
                 <StyledTopBarDiv data-tid="container">
-                    <StyledRow>
+                    <StyledRow>{ folderFiles ? <>
                         <div className="col-sm-4"><b>Tag:</b> {tag}</div>
-                        <div className="col-sm-2">{
-                            noRSMLInFolder ? <b>No RSML found in Folder</b> : <>
-                                <b>Open Image: </b> 
-                                {Object.keys(folderFiles).indexOf(tag) + 1} of {Object.keys(folderFiles).length}
-                                <TooltipOverlay  component={ props => <StyledIcon
-                                        className={"fas fa-info-circle"}
-                                        {...props}
-                                    />} 
-                                    text={"Any images with missing RSML vice versa will be skipped."} // Temporary solution
-                                    placement={"bottom"}
-                                /> 
-                            </>
-                        }</div>
+                        <div className="col-sm-2">{message}
+                            <TooltipOverlay component={ props => <StyledIcon
+                                    className={"fas fa-info-circle"}
+                                    {...props}
+                                />} 
+                                text={"Any images with missing RSML vice versa will be skipped."} // Temporary solution
+                                placement={"bottom"}
+                            /> 
+                        </div>
                         <div className="col-sm-3"><b>Analysis Date:</b> {date}</div>
                         <div className="col-sm-3"><b>Number of Plants:</b> {plants}</div>
-                    </StyledRow>
+                    </> : <><b>No Images Loaded:</b>&nbsp;{"Please select a folder on the left hand side to open."}</>}</StyledRow>
                 </StyledTopBarDiv>
                 <StyledTopBarHR/>
                 <StyledTopBarDiv className="d-inline-flex container" data-tid="container" style={{paddingTop: '0', minWidth: '100%'}}>
@@ -56,8 +59,8 @@ export default class TopBar extends Component {
                             </div>
                         </div>
                         <div className="col-4" style={{textAlign: "center"}}>
-                            <LeftButton click={buttonHandler}/>
-                            <RightButton click={buttonHandler}/>
+                            <LeftButton click={buttonHandler} disabled={!folderFiles}/>
+                            <RightButton click={buttonHandler} disabled={!folderFiles}/>
                         </div>
                         <div className="col-4" style={{textAlign: "right"}}>
                             <ResetChangesButton />